+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes:
 
 * Now set a default Discord 'Post as' username to help speed up setup. You should still set your own though.
+* Reorganized the preferences display into tabs, sections and preferences.
 
 ### Bug Fixes: